Extract function classes to core package.
