+++ cli/src/main/scala/org/bdgenomics/cannoli/cli/SnpEff.scala
@@ -28,73 +28,12 @@ import org.bdgenomics.adam.rdd.variant.{
   VCFOutFormatter
 }
 import org.bdgenomics.adam.sql.{ VariantContext => VariantContextProduct }
+import org.bdgenomics.cannoli.{ SnpEff => SnpEffFn, SnpEffArgs => SnpEffFnArgs }
 import org.bdgenomics.cannoli.builder.CommandBuilders
 import org.bdgenomics.utils.cli._
 import org.bdgenomics.utils.misc.Logging
 import org.kohsuke.args4j.{ Argument, Option => Args4jOption }
 import scala.collection.JavaConversions._
-
-/**
- * SnpEff function arguments.
- */
-class SnpEffFnArgs extends Args4jBase {
-  @Args4jOption(required = false, name = "-executable", usage = "Path to the SnpEff executable. Defaults to snpEff.")
-  var executable: String = "snpEff"
-
-  @Args4jOption(required = false, name = "-image", usage = "Container image to use. Defaults to quay.io/biocontainers/snpeff:4.3.1t--0.")
-  var image: String = "quay.io/biocontainers/snpeff:4.3.1t--0"
-
-  @Args4jOption(required = false, name = "-sudo", usage = "Run via sudo.")
-  var sudo: Boolean = false
-
-  @Args4jOption(required = false, name = "-use_docker", usage = "If true, uses Docker to launch SnpEff.")
-  var useDocker: Boolean = false
-
-  @Args4jOption(required = false, name = "-use_singularity", usage = "If true, uses Singularity to launch SnpEff.")
-  var useSingularity: Boolean = false
-
-  @Args4jOption(required = false, name = "-database", usage = "SnpEff database name. Defaults to GRCh38.86.")
-  var database: String = "GRCh38.86"
-}
-
-/**
- * SnpEff wrapper as a function VariantContextRDD &rarr; VariantContextRDD,
- * for use in cannoli-shell or notebooks.
- *
- * @param args SnpEff function arguments.
- * @param stringency Validation stringency. Defaults to ValidationStringency.LENIENT.
- * @param sc Spark context.
- */
-class SnpEffFn(
-    val args: SnpEffFnArgs,
-    val stringency: ValidationStringency = ValidationStringency.LENIENT,
-    sc: SparkContext) extends CannoliFn[VariantContextRDD, VariantContextRDD](sc) with Logging {
-
-  override def apply(variantContexts: VariantContextRDD): VariantContextRDD = {
-
-    var builder = CommandBuilders.create(args.useDocker, args.useSingularity)
-      .setExecutable(args.executable)
-      .add("-download")
-      .add(args.database)
-
-    if (args.useDocker || args.useSingularity) {
-      builder
-        .setImage(args.image)
-        .setSudo(args.sudo)
-    }
-
-    log.info("Piping {} to snpEff with command: {} files: {}",
-      variantContexts, builder.build(), builder.getFiles())
-
-    implicit val tFormatter = VCFInFormatter
-    implicit val uFormatter = new VCFOutFormatter(sc.hadoopConfiguration, stringency)
-
-    variantContexts.pipe[VariantContext, VariantContextProduct, VariantContextRDD, VCFInFormatter](
-      cmd = builder.build(),
-      files = builder.getFiles()
-    )
-  }
-}
 
 object SnpEff extends BDGCommandCompanion {
   val commandName = "snpEff"
